Issue 169.
Changed interface SemanticTagProvider to use Token as param. instead of a (lemma) String. In the more general case, e.g., when the SemanticTagProvider is implemented as a
Uby-specific resource, the implemented resource needs to have access to particular features of the token, such as lemma *and* POS.

New interface method:
String getSemanticTag(Token token)