#954 Changes to LXF DKPRO that makes the prototype working.
The major difference is in the DKPro2Lxf. The LXF needs that each layer
is created by a tool with corresponding name. If the layer of annotation
is new the DKPro2Lxf will use the id of the tool in this map to create
the nodes corresponding to this layer. The tool id is need for the
origin part of the node. The appropriate test are added that reflect
this major change. Mainly the ref.lxf files are generated with the
default dkpro tool as origin.